refactor - refactoring staticData api folder & add queryKey
